+++ src/components/Query/QueryFormBase.js
@@ -1,3 +1,5 @@
+import { mapVuexAccessors } from '../../helpers/store'
+
 export default {
   name: 'query-form-base',
   data () {
@@ -8,37 +10,6 @@ export default {
     }
   },
   computed: {
-    host: {
-      get () {
-        return this.$store.state.query.host
-      },
-      set (host) {
-        this.$store.commit('setQueryHost', host)
-      }
-    },
-    method: {
-      get () {
-        return this.$store.state.query.method
-      },
-      set (method) {
-        this.$store.commit('setQueryMethod', method)
-      }
-    },
-    stringifiedParams: {
-      get () {
-        return this.$store.state.query.stringifiedParams
-      },
-      set (params) {
-        this.$store.commit('setQueryStringifiedParams', params)
-      }
-    },
-    stringifiedHeaders: {
-      get () {
-        return this.$store.state.query.stringifiedHeaders
-      },
-      set (params) {
-        this.$store.commit('setQueryStringifiedHeaders', params)
-      }
-    }
+    ...mapVuexAccessors('query', ['host', 'method', 'stringifiedParams', 'stringifiedHeaders'])
   }
 }