+++ src/components/Search/ResultsTable.vue
@@ -4,13 +4,13 @@
       <div class="clearfix">
         <v-flex right d-inline-flex>
           <v-text-field append-icon="search"
-                        @keyup.esc="searchFilter = ''"
+                        @keyup.esc="filter = ''"
                         label="Filter results..."
                         messages="Filter via 'column:query'"
                         name="filter"
                         id="filter"
                         class="mt-0"
-                        v-model="searchFilter"></v-text-field>
+                        v-model="filter"></v-text-field>
         </v-flex>
       </div>
     </v-card-text>
@@ -19,16 +19,16 @@
                   :headers="headers"
                   :items="flattenedHits"
                   :loading="loading"
-                  :search="searchFilter"
+                  :search="filter"
                   :custom-filter="callFuzzyTableFilter"
-                  :pagination.sync="searchPagination"
+                  :pagination.sync="pagination"
                   class="table--condensed fixed-header">
       <template slot="items" slot-scope="item">
         <tr @click="openDocument(item.item)" class="tr--clickable">
-          <td>{{ item.item._index }}</td>
+          <td v-if="showIndex">{{ item.item._index }}</td>
+          <td v-if="showScore">{{ item.item._score}}</td>
           <td>{{ item.item._id}}</td>
           <td>{{ item.item._type}}</td>
-          <td>{{ item.item._score}}</td>
           <td v-for="key in keys" :key="item.item._index + '_' + key">{{item.item[key]}}</td>
         </tr>
       </template>
@@ -45,10 +45,10 @@
 <script>
   import { objectArrayUniqueKeys } from '../../helpers/utilities'
   import { fuzzyTableFilter } from '../../helpers/filters'
-  import FixedHeaderTable from '@/mixins/FixedHeaderTable'
+  import FixedTableHeader from '@/mixins/FixedTableHeader'
   import { DEFAULT_ROWS_PER_PAGE } from '../../consts'
-
-  const DEFAULT_KEYS = ['_index', '_id', '_type', '_score']
+  import { mapVuexAccessors } from '../../helpers/store'
+  import { mapState } from 'vuex'
 
   export default {
     name: 'ResultsTable',
@@ -64,8 +64,16 @@
       }
     },
     computed: {
+      defaultKeys () {
+        let keys = []
+        if (this.showIndex) keys.push('_index')
+        if (this.showScore) keys.push('_score')
+        keys.push('_id')
+        keys.push('_type')
+        return keys
+      },
       headers () {
-        let defaultKeyHeaders = DEFAULT_KEYS.map(value => ({text: value, value: value}))
+        let defaultKeyHeaders = this.defaultKeys.map(value => ({text: value, value: value}))
         return defaultKeyHeaders.concat(this.keys.map(value => ({text: value, value: value})))
       },
       keys () {
@@ -78,22 +86,8 @@
           return hit
         })
       },
-      searchFilter: {
-        get () {
-          return this.$store.state.search.filter
-        },
-        set (filter) {
-          this.$store.commit('setSearchFilter', filter)
-        }
-      },
-      searchPagination: {
-        get () {
-          return this.$store.state.search.pagination
-        },
-        set (pagination) {
-          this.$store.commit('setSearchPagination', pagination)
-        }
-      }
+      ...mapVuexAccessors('search', ['filter', 'pagination']),
+      ...mapState('search', ['showIndex', 'showScore'])
     },
     methods: {
       openDocument (item) {
@@ -107,13 +101,13 @@
       }
     },
     mounted () {
-      this.fixedHeaderTableOnMount()
+      this.fixedTableHeaderOnEnable()
     },
     beforeDestroy () {
-      this.fixedHeaderTableOnBeforeDestroy()
+      this.fixedTableHeaderOnDisable()
     },
     mixins: [
-      FixedHeaderTable
+      FixedTableHeader
     ]
   }
 </script>